python/requirements.txt
quart
quart_schema
pydantic
hypercorn
adafruit-circuitpython-dht
adafruit-blinka
RPi.GPIO
w1thermsensor
aiorwlock